Refactor clause subsumption.
